--- docs/api/ocsp.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.ocsp`
------------------------
-
-.. automodule:: certbot.ocsp
-   :members:
